Move test as per CR comments
